refactor: Update extractNumberedElements to use htmlparser2 for parsing HTML content
