+++ web/app/features/prepare-html-for-translate/utils/addNumbersToContent.ts
