+++ web/app/features/translate/constants.ts
