+++ web/app/features/translate/types.ts
@@ -1,0 +1,4 @@
+export type NumberedElement = {
+	number: number;
+	text: string;
+};
