+++ web/app/routes/translator/functions/queries.server.ts
