+++ web/app/routes/translator/route.tsx
@@ -4,8 +4,8 @@ import { useRevalidator } from "@remix-run/react";
 import { useEffect } from "react";
 import { typedjson, useTypedLoaderData } from "remix-typedjson";
 import { Header } from "~/components/Header";
-import { getTranslateUserQueue } from "~/feature/translate/translate-user-queue";
-import { getDbUser } from "~/libs/db/user.server";
+import { getTranslateUserQueue } from "~/features/translate/translate-user-queue";
+import { getDbUser } from "./functions/queries.server";
 import { authenticator } from "~/utils/auth.server";
 import { normalizeAndSanitizeUrl } from "~/utils/normalize-and-sanitize-url.server";
 import { getTargetLanguage } from "~/utils/target-language.server";
