+++ web/app/routes/translator/types.ts
@@ -41,7 +41,3 @@ export type PageVersionTranslationInfoItem = z.infer<
 	typeof PageVersionTranslationInfoSchema
 >;
 
-export type NumberedElement = {
-	number: number;
-	text: string;
-};
